Allow empty log messages.
These were deliberately being thrown away, but for no apparent reason.

The log message will eventually contain useful information, like the
source location, time, and thread, and so there's value in having a log
message emitted even if the "message" is empty.

This means that client code can do log.debug("") and still see when that
line is reached, or log.error("\(serverError)") and still see that
something bad happened, even if the server returned nothing.