Add optional message filter to BaseDestination
It might be desirable to further filter messages based upon the message content itself.
Since filtering by message content first requires that the message be resolved to a String, thereby
defeating the purpose of the @autoclosure, the minLevelFilters are first applied. If the message is
not filtered by the level filtering using level, path and function, then the message is resolved
(once instead of once for each destination) and if there are any message filters in place, those
are invoked to see if any message filters reject the message.

This change provides additional flexibility for content-based filtering where level, path and function filtering
is not enough.

Resolving the message once for each destination introduces the remote possibility of a bug. If the expression
changes each time it is run, the message could be different for each destination. This change resolves the message
once and only once per log request regardless of how many destinations there are.

Only if the message gets through all message filters is it sent to the destination.

Additional tests have been written to verify.