Add content filtering
This PR allows Content Filters to be added to a destination to further
control what gets sent to the destination. A ContentFilter, when applied, answers
true or false. Answering true allows the message to be logged. If multiple content
filters have been added to a destination, ANY filter answering true will allow the message
to be logged even if others answer false.