Add new filtering implementation and deprecated minLevelFilters
There are now multiple individual and flexible filters that can be
added/removed to a destination to better control what gets logged
to the destination. You can add any combination of Path, Function and/or
Message filters using a flexible comparison mechanism. Logged messages
can use startsWith, contains, endsWith or equals filters and each can
be case sensitive or not.